requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu126

accelerate==1.10.1
aio-pika==9.4.1
aiokafka==0.12.0
albumentations==1.4.1
amqp==5.2.0
datasets==4.1.1
einops==0.8.1
faiss-cpu==1.12.0
gdown==5.2.0
gradio==5.46.1
hf-xet==1.1.10
huggingface-hub==0.35.0
kafka-python==2.2.15
loguru==0.7.3
matplotlib==3.10.6
monai==1.5.1
nibabel==5.3.2
numpy==2.2.6
nvitop==1.5.3
onnx==1.19.0
onnxruntime-gpu==1.22.0
opencv-python-headless==4.12.0.88
pandas==2.3.2
pillow==11.3.0
pydicom==3.0.1
pynrrd==1.1.3
python-dotenv==1.1.1
pytorch-lightning==2.5.5
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentencepiece==0.2.1
simpleitk==2.5.2
sympy==1.13.1
tokenizers==0.13.3
torch==2.6.0
torchio==0.20.22
torchvision==0.21.0
transformers==4.30.2
vector-quantize-pytorch==1.1.2